% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_preds.R
\name{combine_preds}
\alias{combine_preds}
\title{Combine predictions from models into one dataframe}
\usage{
combine_preds(rf_preds, boost_preds, glm_preds, lasso_preds, ridge_preds)
}
\description{
Combine predictions from models into one dataframe
}
