% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_folds.R
\name{create_custom_folds}
\alias{create_custom_folds}
\title{Create folds for cross validation}
\usage{
create_custom_folds(x, k, seed)
}
\arguments{
\item{k}{number of folds}

\item{seed}{seed for random sample}

\item{training}{data frame for training model}
}
\description{
Create folds for cross validation
}
