% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create a summary table of participant characteristics}
\usage{
create_summary_table(combined)
}
\arguments{
\item{combined}{dataframe used for modelling}
}
\description{
Create a summary table of participant characteristics
}
