% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for outcome}
\usage{
format_data(resp_data, durations, lower_trigger, upper_trigger)
}
\description{
The predictors in this model will be total dose of
sedation, time that sedation was administered, duration and time
since previous apneic event, and (non-missing) patient
demographics.
}
