% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_train_id.R
\name{make_train_id}
\alias{make_train_id}
\title{Adds a variable to allocate to training or testing}
\usage{
make_train_id(x, seed, frac)
}
\description{
Adds a variable to allocate to training or testing
}
